backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
pymysql
python-dotenv
